requirements.txt
python-pptx==0.6.18
lxml==4.1.0